+++ .eslintrc.js
@@ -102,9 +102,9 @@ module.exports = {
       },
     },
     {
-      files: ['src/XRP/default-xrp-client.ts'],
+      files: ['src/BRT/default-xrp-client.ts'],
       rules: {
-        // This is actually a good rule to have enabled, but for the XRPClient, we define a helper error message class in the same file
+        // This is actually a good rule to have enabled, but for the BRTClient, we define a helper error message class in the same file
         'max-classes-per-file': 'off',
       },
     },