+++ src/types/amount.ts
@@ -15,7 +15,7 @@ const MIN_IOU_EXPONENT = -96;
 const MAX_IOU_EXPONENT = 80;
 const MAX_IOU_PRECISION = 16;
 const MAX_DROPS = new Decimal("1e17");
-const MIN_XRP = new Decimal("1e-6");
+const MIN_BRT = new Decimal("1e-6");
 const mask = bigInt(0x00000000ffffffff);
 
 /**
@@ -135,8 +135,8 @@ class Amount extends SerializedType {
    * @returns An Amount object
    */
   static fromParser(parser: BinaryParser): Amount {
-    const isXRP = parser.peek() & 0x80;
-    const numBytes = isXRP ? 48 : 8;
+    const isBRT = parser.peek() & 0x80;
+    const numBytes = isBRT ? 48 : 8;
     return new Amount(parser.read(numBytes));
   }
 
@@ -186,9 +186,9 @@ class Amount extends SerializedType {
   }
 
   /**
-   * Validate XRP amount
+   * Validate BRT amount
    *
-   * @param amount String representing XRP amount
+   * @param amount String representing BRT amount
    * @returns void, but will throw if invalid amount
    */
   private static assertXrpIsValid(amount: string): void {
@@ -198,7 +198,7 @@ class Amount extends SerializedType {
 
     const decimal = new Decimal(amount);
     if (!decimal.isZero()) {
-      if (decimal.lt(MIN_XRP) || decimal.gt(MAX_DROPS)) {
+      if (decimal.lt(MIN_BRT) || decimal.gt(MAX_DROPS)) {
         throw new Error(`${amount.toString()} is an illegal amount`);
       }
     }
@@ -244,9 +244,9 @@ class Amount extends SerializedType {
   }
 
   /**
-   * Test if this amount is in units of Native Currency(XRP)
+   * Test if this amount is in units of Native Currency(BRT)
    *
-   * @returns true if Native (XRP)
+   * @returns true if Native (BRT)
    */
   private isNative(): boolean {
     return (this.bytes[0] & 0x80) === 0;